some fixes, add readme, add tests

+++ src/main/java/com/redcare/challenge/service/DefaultScoringService.java
@@ -18,7 +18,7 @@ public class DefaultScoringService implements ScoringService {
     private static final int RECENT_DAYS_THRESHOLD = 30;
 
     @Override
-    public int calculateScore(int stars, int forks, Instant updatedAt) {
+    public double calculateScore(int stars, int forks, Instant updatedAt) {
         int score = stars + forks * 2;
 
         long daysOld = ChronoUnit.DAYS.between(updatedAt, Instant.now());
