+++ src/test/java/com/redcare/challenge/service/DefaultScoringServiceTest.java
@@ -23,22 +23,22 @@ class DefaultScoringServiceTest {
         Instant tenDaysAgo = Instant.now().minus(30, ChronoUnit.DAYS);
         Instant fortyDaysAgo = Instant.now().minus(31, ChronoUnit.DAYS);
 
-        int highScore = scoringService.calculateScore(100, 50, tenDaysAgo);
-        int lowScore = scoringService.calculateScore(100, 50, fortyDaysAgo);
+        double highScore = scoringService.calculateScore(100, 50, tenDaysAgo);
+        double lowScore = scoringService.calculateScore(100, 50, fortyDaysAgo);
 
         assertAll("score calculation",
                 () -> assertEquals(
                         210,
                         highScore,
                         () -> String.format(
-                                "Repositories updated exactly 30 days ago should receive the recency bonus and have a score of 210 with 100 Stars and 50 forks, but was %d",
+                                "Repositories updated exactly 30 days ago should receive the recency bonus and have a score of 210 with 100 Stars and 50 forks, but was %.2f",
                                 highScore)
                 ),
                 () -> assertEquals(
                         200,
                         lowScore,
                         () -> String.format(
-                                "Repositories updated 31 days ago should not receive the recency bonus and have a score of 200 with 100 Stars and 50 forks, but was %d",
+                                "Repositories updated 31 days ago should not receive the recency bonus and have a score of 200 with 100 Stars and 50 forks, but was %.2f",
                                 lowScore)
                 )
         );
